Fixed LAN connections outside of host
Users can now connect to other devices on the wifi. Firewall has to be turned off for this to work, and the user needs the unity project open in the Unity editor to enable the start button. After some polish this weekend, I'll get all the setup possible within the game menus.